Backend/requirements.txt
fastapi
uvicorn
pikachu-chem
